ft: move files around
