+++ projects/ngx-semantic/ng-package.json
@@ -2,6 +2,6 @@
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
   "dest": "../../dist/ngx-semantic",
   "lib": {
-    "entryFile": "./ngx-semantic.module.ts"
+    "entryFile": "./src/ngx-semantic.module.ts"
   }
 }
