+++ projects/ngx-semantic/src/common/enums.ts
