+++ projects/ngx-semantic/src/common/public-api.ts
@@ -1,0 +1,5 @@
+/**
+ * Created by bolor on 4/22/2020
+ */
+
+export * from './enums';
