+++ projects/ngx-semantic/src/ngx-semantic.module.ts
@@ -2,6 +2,7 @@ import {NgModule} from '@angular/core';
 import {SuiButtonModule} from './button';
 
 export * from './button';
+export * from './common';
 
 @NgModule({
   declarations: [],
