+++ projects/ngx-semantic/src/public-api.ts
@@ -1,0 +1,3 @@
+/**
+ * Created by bolor on 4/22/2020
+ */
